# api/openapi.yaml
openapi: 3.0.3
info:
  title: ESNODE-Runtime API
  version: 0.1.0
paths:
...
  /esnode/v1/models/load:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModelSpec" }
      responses:
        "200":
          description: Loaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_handle: { type: string }
  /esnode/v1/models/unload:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model_handle: { type: string }
      responses:
        "200":
          description: Unloaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
...
  /esnode/v1/infer:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InferRequest" }
      responses:
        "200":
          description: Inference response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InferResponse" }
  /esnode/v1/infer/stream:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InferRequest" }
      responses:
        "200":
          description: Inference stream
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                data: {"kind":"start","request_id":"req-1","model":"demo-gguf","metadata":{}}

                data: {"kind":"delta","request_id":"req-1","delta_text":"hello"}

                data: {"kind":"end","request_id":"req-1","usage":{"prompt_tokens":0,"completion_tokens":0,"total_tokens":0}}
  /v1/chat/completions:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OpenAIChatCompletionRequest" }
      responses:
        "200":
          description: OpenAI-style response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OpenAIChatCompletionResponse" }
            text/event-stream:
              schema:
                type: string
              example: |
                data: {"id":"cmpl-stub","object":"chat.completion.chunk","created":0,"model":"local-gguf","choices":[{"index":0,"delta":{"role":"assistant"},"finish_reason":null}]}

                data: {"id":"cmpl-stub","object":"chat.completion.chunk","created":0,"model":"local-gguf","choices":[{"index":0,"delta":{"content":"hello"},"finish_reason":null}]}

                data: [DONE]
...
components:
  schemas:
    TokenUsage:
      type: object
      properties:
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }
    ModelSpec:
      type: object
      required: [name, version, source, format, backend, backend_settings]
      properties:
        name: { type: string }
        version: { type: string }
        source:
          type: object
          required: [kind]
          properties:
            kind: { type: string, enum: [local_path, http, registry] }
            path: { type: string }
            url: { type: string }
            sha256: { type: string }
            registry_name: { type: string }
            digest: { type: string }
        format: { type: string, enum: [auto, onnx, gguf, torchscript, huggingface] }
        backend: { type: string, description: "auto|onnxrt|llamacpp|torch" }
        backend_settings: { type: object, additionalProperties: true }
        labels: { type: object, additionalProperties: { type: string } }
    InferRequest:
      type: object
      required: [request_id, model, input, params]
      properties:
        request_id: { type: string }
        model: { type: string }
        input:
          oneOf:
            - $ref: "#/components/schemas/ChatInput"
            - $ref: "#/components/schemas/CompletionInput"
        params:
          type: object
          properties:
            max_tokens: { type: integer }
            temperature: { type: number }
            top_p: { type: number }
            top_k: { type: integer }
            stop: { type: array, items: { type: string } }
            seed: { type: integer }
            stream: { type: boolean }
            timeout_ms: { type: integer }
        metadata: { type: object, additionalProperties: { type: string } }
    ChatInput:
      type: object
      required: [type, messages]
      properties:
        type: { type: string, enum: [chat] }
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role: { type: string }
              content: { type: string }
              name: { type: string }
    CompletionInput:
      type: object
      required: [type, prompt]
      properties:
        type: { type: string, enum: [completion] }
        prompt: { type: string }
    InferResponse:
      type: object
      required: [request_id, model, output]
      properties:
        request_id: { type: string }
        model: { type: string }
        output:
          oneOf:
            - type: object
              properties:
                type: { type: string, enum: [chat] }
                message:
                  type: object
                  properties:
                    role: { type: string }
                    content: { type: string }
            - type: object
              properties:
                type: { type: string, enum: [completion] }
                text: { type: string }
        usage:
          $ref: "#/components/schemas/TokenUsage"
    OpenAIChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string }
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role: { type: string }
              content: { type: string }
        stream: { type: boolean }
        max_tokens: { type: integer }
        temperature: { type: number }
        top_p: { type: number }
    OpenAIChatCompletionResponse:
      type: object
      properties:
        id: { type: string }
        object: { type: string }
        created: { type: integer }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message:
                type: object
                properties:
                  role: { type: string }
                  content: { type: string }
              finish_reason: { type: string }
        usage:
          $ref: "#/components/schemas/TokenUsage"
...